+++ packages/components/src/components/select/select.styles.ts
@@ -162,6 +162,10 @@ export default css`
     margin-inline-end: var(--syn-input-spacing-small);
   }
 
+  .select--small.select--multiple .select__prefix::slotted(*) {
+    margin-inline-start: var(--syn-input-spacing-small);
+  }
+
   .select--small.select--multiple:not(.select--placeholder-visible) .select__combobox {
     padding-block: 2px;
     padding-inline-start: 0;
@@ -185,6 +189,10 @@ export default css`
 
   .select--medium .select__prefix::slotted(*) {
     margin-inline-end: var(--syn-input-spacing-medium);
+  }
+
+  .select--medium.select--multiple .select__prefix::slotted(*) {
+    margin-inline-start: var(--syn-input-spacing-medium);
   }
 
   .select--medium.select--multiple:not(.select--placeholder-visible) .select__combobox {
@@ -212,6 +220,10 @@ export default css`
     margin-inline-end: var(--syn-input-spacing-large);
   }
 
+  .select--large.select--multiple .select__prefix::slotted(*) {
+    margin-inline-start: var(--syn-input-spacing-large);
+  }
+
   .select--large.select--multiple:not(.select--placeholder-visible) .select__combobox {
     padding-inline-start: 0;
     padding-block: 4px;
@@ -219,20 +231,17 @@ export default css`
 
   .select--large .select__tags {
     gap: 4px;
-  }/* Prefix */
-  .select__prefix {
+  }/* Prefix and Suffix */
+  .select__prefix,
+  .select__suffix {
     flex: 0;
     display: inline-flex;
     align-items: center;
     color: var(--syn-input-placeholder-color);
   }
 
-  /* Suffix */
-  .select__suffix {
-    flex: 0;
-    display: inline-flex;
-    align-items: center;
-    color: var(--syn-input-placeholder-color);
+  .select__suffix::slotted(*) {
+    margin-inline-start: var(--syn-spacing-small);
   }
 
   /* Clear button */