+++ packages/components/src/events/syn-close.ts
@@ -6,6 +6,9 @@
  */
 
 /* eslint-disable */
+/**
+ * Synergy custom event
+ */
 export type SynCloseEvent = CustomEvent<Record<PropertyKey, never>>;
 
 declare global {