+++ packages/components/src/events/syn-initial-focus.ts
@@ -6,6 +6,9 @@
  */
 
 /* eslint-disable */
+/**
+ * Synergy custom event
+ */
 export type SynInitialFocusEvent = CustomEvent<Record<PropertyKey, never>>;
 
 declare global {